refactor: extract subagent tracking logic into standalone modules
Move subagent state management and persistence out of the monolithic
openclawRuntimeAdapter.ts and coworkStore.ts into dedicated files:
- SubagentRunStore: DB CRUD for subagent_runs table
- SubagentTracker: in-memory state, lifecycle hooks, history fetching
